Code review feedback: test npg_irods.functions rather than npg_irods.cli.publish_directory.main with mocking - clearer
